requirements.txt
amqp==2.6.1
asgiref==3.3.4
backcall==0.2.0
billiard==3.6.4.0
celery==4.4.6
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.9
Django==3.1.4
django-cors-headers==3.5.0
django-redis==4.12.1
djangorestframework==3.12.2
future==0.18.2
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
kombu==4.6.11
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.5
XlsxWriter==1.4.3
